// lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        Build: {
            hero_build: components["schemas"]["BuildHero"];
            /** Format: int32 */
            num_favorites?: number | null;
            /** Format: int32 */
            num_ignores?: number | null;
            /** Format: int32 */
            num_reports?: number | null;
            /** Format: int32 */
            num_weekly_favorites?: number | null;
            /** Format: int32 */
            rollup_category?: number | null;
        };
// ...
        HeroBuildStats: {
            /**
             * Format: int64
             * @description The ID of the hero build. The `hero_build_id` is the first build the player had selected when the game started.
             */
            hero_build_id: number;
            /**
             * Format: int32
             * @description The ID of the hero. See more: <https://api.deadlock-api.com/v1/assets/heroes>
             */
            hero_id: number;
            /**
             * Format: int64
             * @description The number of losses with this build.
             */
            losses: number;
            /**
             * Format: int64
             * @description The total number of matches played with this build (`wins + losses`).
             */
            matches: number;
            /**
             * Format: int64
             * @description The number of unique players who used this build.
             */
            players: number;
            /**
             * Format: int64
             * @description The number of wins with this build.
             */
            wins: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}